Move match and allocate service
