+++ test/services/bill-runs/two-part-tariff/process-bill-run.service.test.js
@@ -14,7 +14,7 @@ const BillRunModel = require('../../../../app/models/bill-run.model.js')
 
 // Things we need to stub
 const HandleErroredBillRunService = require('../../../../app/services/bill-runs/handle-errored-bill-run.service.js')
-const MatchAndAllocateService = require('../../../../app/services/bill-runs/two-part-tariff/match-and-allocate.service.js')
+const MatchAndAllocateService = require('../../../../app/services/bill-runs/match/match-and-allocate.service.js')
 
 // Thing under test
 const TwoPartTariffProcessBillRunService = require('../../../../app/services/bill-runs/two-part-tariff/process-bill-run.service.js')
